finished primitive-to-Integer assignment

+++ malachite-native/src/integer/conversion/assign_i64.rs
@@ -1,7 +1,7 @@
 use integer::Integer;
 use traits::Assign;
 
-/// Assigns an `i32` to an `Integer`.
+/// Assigns an `i64` to an `Integer`.
 ///
 /// Time: worst case O(1)
 ///
@@ -13,12 +13,12 @@ use traits::Assign;
 /// use malachite_native::traits::Assign;
 ///
 /// let mut x = Integer::from(456);
-/// x.assign(-123);
+/// x.assign(-123i64);
 /// assert_eq!(x.to_string(), "-123");
 /// ```
-impl Assign<i32> for Integer {
-    fn assign(&mut self, other: i32) {
+impl Assign<i64> for Integer {
+    fn assign(&mut self, other: i64) {
         self.sign = other >= 0;
-        self.abs.assign(other.abs() as u32);
+        self.abs.assign(other.abs() as u64);
     }
 }
